requirements.txt
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
et_xmlfile==2.0.0
Flask==3.1.0
git-filter-repo==2.47.0
google-auth==2.40.1
google-auth-oauthlib==1.2.2
gspread==6.2.0
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==5.4.0
MarkupSafe==3.0.2
numpy==2.2.5
oauth2client==4.1.3
oauthlib==3.2.2
openpyxl==3.1.5
packaging==25.0
pandas==2.2.3
pillow==11.2.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyee==11.1.1
pyparsing==3.2.3
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.0
python-utils==3.9.1
pytz==2025.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
schedule==1.2.2
setuptools==80.3.1
six==1.17.0
slack_bolt==1.23.0
slack_sdk==3.35.0
slackeventsapi==3.0.3
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
Werkzeug==3.1.3
